% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM.R
\name{gcm_test}
\alias{gcm_test}
\title{Calculate the covariance between the calculated GCMs from two different GCM tests}
\usage{
gcm_test(
  X = NULL,
  Y = NULL,
  Z = NULL,
  alpha = 0.05,
  regr.method = "xgboost",
  regr.pars = list(),
  plot.residuals = FALSE,
  nsim = 499L,
  resid.XonZ = NULL,
  resid.YonZ = NULL
)
}
\arguments{
\item{X}{TODO}

\item{Y}{TODO}

\item{Z}{TODO}

\item{alpha}{TODO}

\item{regr.method}{TODO}

\item{regr.pars}{TODO}

\item{plot.residuals}{TODO}

\item{nsim}{TODO}

\item{resid.XonZ}{TODO}

\item{resid.YonZ}{TODO}
}
\value{
calculated covariance
}
\description{
Calculate the covariance between the calculated GCMs from two different GCM tests
}
